debugging and adding gradient_map initialization to main logic

+++ src/utils/gradient_cell.rs
@@ -1,9 +1,9 @@
 use hlt::position::Position;
 
 pub struct GradientCell {
-    pub position: Position
-    pub value: f64
-    pub nearby_ship_count: i8
+    pub position: Position,
+    pub value: usize,
+    pub nearby_ship_count: i8,
 }
